# docs/swagger.yaml
basePath: /api
definitions:
  ds.Class:
    properties:
      color:
        type: string
      examples:
        type: string
      id:
        type: integer
      image:
        type: string
      isDeleted:
        type: boolean
      massRequestToClass:
        items:
          $ref: '#/definitions/ds.MassRequestToClass'
        type: array
      name:
        type: string
      spectre:
        type: string
      temperature:
        type: integer
    type: object
  ds.MassRequest:
    properties:
      closedAt:
        type: string
      createdAt:
        type: string
      formedAt:
        type: string
      id:
        type: integer
      massRequestToClass:
        items:
          $ref: '#/definitions/ds.MassRequestToClass'
        type: array
      moderator:
        $ref: '#/definitions/ds.User'
      moderatorId:
        type: integer
      status:
        description: 1 - draft, 2 - deleted, 3 - pending, 4 - resolved, 5 - rejected
        type: integer
      user:
        $ref: '#/definitions/ds.User'
      userID:
        type: integer
    type: object
  ds.MassRequestToClass:
    properties:
      calcRequest:
        $ref: '#/definitions/ds.MassRequest'
      class:
        $ref: '#/definitions/ds.Class'
      classID:
        type: integer
      luminosity:
        type: integer
      mass:
        type: integer
      requestID:
        type: integer
    type: object
  ds.User:
    properties:
      id:
        type: integer
      isMod:
        type: boolean
      passHash:
        type: string
      passSalt:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для расчета масс звезд в заявках
  title: Stellar Mass Calculation API
  version: "1.0"
paths:
  /mass-requests:
    get:
      consumes:
      - application/json
      description: Get mass requests with optional filters
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: integer
      - description: Filter from date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Filter to date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.MassRequest'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get mass requests
      tags:
      - mass-requests
    post:
      consumes:
      - application/json
      description: Create a new draft mass request
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create mass request
      tags:
      - mass-requests
  /mass-requests/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a draft mass request
      parameters:
      - description: Mass Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete mass request
      tags:
      - mass-requests
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific mass request
      parameters:
      - description: Mass Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.MassRequest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get mass request by ID
      tags:
      - mass-requests
    put:
      consumes:
      - application/json
      description: Update fields of a draft mass request
      parameters:
      - description: Mass Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update mass request
      tags:
      - mass-requests
...
  /mass-requests/{id}/complete:
    put:
      consumes:
      - application/json
      description: Complete or reject a mass request with calculations (moderator
        only)
      parameters:
      - description: Mass Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Action data
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Complete mass request
      tags:
      - mass-requests
  /mass-requests/{id}/form:
    put:
      consumes:
      - application/json
      description: Submit draft request for moderation
      parameters:
      - description: Mass Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Form mass request
      tags:
      - mass-requests
...
  /mass-requests/star-calculation:
    get:
      consumes:
      - application/json
      description: Get draft request ID and item count for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get star calculation cart info
      tags:
      - mass-requests
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"